Get rid of authprovider, simplify login/logout
